Extract tests/debugging/gdb.py
